refactor: Move __init__ to base class
